backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.1
starlette==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
requests==2.31.0